+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Add reflection for Component and Default on PointLight2D and AmbientLight (#50).
+
 ## [0.6.0] - 2025-04-29
 
 ### Changed